requirements.txt
# SKILLens v2.0 — Requirements
# ──────────────────────────────────────────────────────────────────────────────
 
# Core UI
streamlit>=1.35.0
 
# NLP
spacy>=3.7.0
nltk>=3.8.1
# After install: python -m spacy download en_core_web_sm
 
# ML / Vectorization
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
 
# Resume Parsing
pdfplumber>=0.11.0
PyMuPDF>=1.24.0        # pymupdf — best for multi-column PDFs
PyPDF2>=3.0.0          # fallback
python-docx>=1.1.0
 
# Visualization
plotly>=5.22.0
matplotlib>=3.8.0
wordcloud>=1.9.3
 
 Pillow>=9.0.0,<10.0.0